requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
python-dotenv
oci
